// src/types/TestCaseStats.ts
export type TestCaseStats = {
  totalPass: number;
  totalFailed: number;
  totalDependent: number;
  totalBlocked: number;
  totalToBeReviewed: number;
  totalIncomplete: number;
  totalDeferred: number;
  totalReadyToRun: number;
  totalUnableToTest: number;
  totalNa: number;
  totalOthers: number;
}